pyproject.toml
[project]
name = "summerschool"
version = "0.1.0"
description = "summerschool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepxde>=1.13.2",
    "dotenv>=0.9.9",
    "jax>=0.6.2",
    "jupyter>=1.1.1",
    "jupytext>=1.16.7",
    "matplotlib>=3.9.3",
    "neuraloperator>=0.3.0",
    "numpy>=2.1.3",
    "pydot>=3.0.4",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "tach>=0.16.3",
    "tensorflow>=2.19.0",
    "torch>=2.6.0",
    "torch-harmonics>=0.7.4",
    "tqdm>=4.67.1",
    "ty>=0.0.1a12",
    "wandb>=0.19.8",
]




[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"sciml" = "."} # "QR1" = "models/QR1/", "QR2" = "models/QR2/"
packages = ["sciml"] # ["QR1", "QR2"]



[tool.setuptools.package-data]
"*" = ["*.py", "*.pyi", "*.c", "*.h", "*.cpp", "*.hpp"]
